cleanup + setting endpoint update
