+++ services/tenant-ui/frontend/src/store/tenantStore.ts
@@ -199,7 +199,7 @@ export const useTenantStore = defineStore('tenant', () => {
       // Get connect_to_endorser and create_public_did
       // from InnkeeperWalletConfig.
       const cConfig = await acapyApi.getHttp(
-        API_PATH.INNKEEPER_CONFIG(tenant.value.tenant_id)
+        API_PATH.TENANT_CONFIG
       );
       console.log(cConfig);
       // Create a DID
